pyproject.toml
[project]
name = "dataGen"
version = "0.1.0"
description = "Generate synthethic data and save as CSV, SQLlite and VectorDB"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "chromadb>=1.0.12",
  "fastapi>=0.115.9",
  "langchain>=0.3.25",
  "langchain-core>=0.3.65",
  "langchain-google-genai>=2.1.5",
  "langgraph>=0.4.8",
  "python-dotenv>=1.1.0",
  "sentence-transformers>=4.1.0",
  "uvicorn>=0.34.3",
]
